feat: verify token

+++ src/app/api/report/route.ts
@@ -7,6 +7,8 @@ import { middleware_1 } from '@/middleware/middleware';
 import { authenticate } from '../utils/decode';
 
 export async function POST(request: any) {
+  let flag = await authenticate(request)
+  if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
   let report: reportType = await request.json();
   await dbConnect();
   await Report.create(report);
@@ -14,16 +16,16 @@ export async function POST(request: any) {
 }
 
 export async function GET(request: any) {
-  
-  
-  authenticate(request)
+  let flag = await authenticate(request)
+  if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
   await dbConnect();
   let reports: reportType[] = await Report.find();
   return NextResponse.json({ reports });
 }
 
 export async function DELETE(request: any) {
-  authenticate(request)
+  let flag = await authenticate(request)
+  if (!flag) return NextResponse.json({ status: 'Error', message: 'Access Denied. Invalid Token.' }, { status: 400 });
   const id = request.nextUrl.searchParams.get('id');
   await dbConnect();
   await Report.findByIdAndDelete(id);
